// src/app/dashboard/default-file-maintenance/df-maintenance-details/temp-set-quality-template/model/setqltytmplt/screen-map.model.ts
import { M79cmsgt } from './m79cmsgt.model';
import { M79ltrst } from './m79ltrst.model';
import { M79msgt } from './m79msgt.model';

/**
 * Model class ScreenMap
 * Path: screenbean/setqltytmpltservice/setqltytmplt
 * Model: com::uhc::aarp::fox::domain::screenbean::setqltytmpltservice::setqltytmplt::ScreenMap
 * Legacy Mapping: RPDMA79T
 */
export class ScreenMap {
  m79tRevpZ = 0;
  m79tAsnaZ = 0;
  m79tAshaZ = 0;
  m79tNasaZ = 0;
  m79tNahaZ = 0;
  m79msgts: M79msgt[] = [];
  m79cmsgts: M79cmsgt[] = [];
  m79ltrsts: M79ltrst[] = [];
  m79headf = '';
  m79smemi = '';
  m79benf = '';
  m79tempi = '';
  m79commi = '';
  m79npayf = '';
  m79msg2f = '';
  m79asnai = '';
  m79revsi = '';
  m79nhosi = '';
  m79headi = '';
  m79nahaf = '';
  m79ashaf = '';
  m79mltrf = '';
  m79pf8kf = '';
  m79msg2i = '';
  m79naspi = '';
  m79sclmf = '';
  m79nasai = '';
  m79decif = '';
  m79verfi = '';
  m79smemf = '';
  m79pdni = '';
  m79exsi = '';
  m79exsf = '';
  m79nahai = '';
  m79beni = '';
  m79tempf = '';
  m79datef = '';
  m79admhi = '';
  m79verif = '';
  m79ashpf = '';
  m79mltri = '';
  m79ndoci = '';
  m79ndocf = '';
  m79ashpi = '';
  m79asnpf = '';
  m79npayi = '';
  m79payf = '';
  m79specf = '';
  m79corpf = '';
  m79corpi = '';
  m79payi = '';
  m79asnaf = '';
  m79asnpi = '';
  m79pdnf = '';
  m79ionsi = '';
  m79payaf = '';
  m79verii = '';
  m79sclmi = '';
  m79ionsf = '';
  m79decii = '';
  m79revpf = '';
  m79speci = '';
  m79nasaf = '';
  m79revdf = '';
  m79ltrsf = '';
  m79revsf = '';
  m79ashai = '';
  m79ltrsi = '';
  m79naspf = '';
  m79revpi = '';
  m79payai = '';
  m79commf = '';
  m79datei = '';
  m79revdi = '';
  m79msg1i = '';
  m79msg1f = '';
  m79nhosf = '';
  m79pf8ki = '';
  m79nahpi = '';
  m79admhf = '';
  m79nahpf = '';
  m79verff = '';
}